# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion PartList",
        "contact": {
            "email": "antony.rodriguez@mimco.com.pe"
        },
        "version": "1"
    },
    "paths": {
...
        "/GestionPartList/public/index.php/regi_proy_pque": {
            "post": {
                "tags": [
                    "Gestion Proyecto Paquete"
                ],
                "summary": "Registrar Proyecto Paquete",
                "operationId": "App\\Http\\Controllers\\ProyectopaqueteController::regi_proy_pque",
                "parameters": [
                    {
                        "name": "intIdProyTarea",
                        "in": "path",
                        "description": "ingresar el id proyecto tarea",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdProy",
                        "in": "path",
                        "description": "ingresar el id  Proyectos",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdTipoProducto",
                        "in": "path",
                        "description": "documento de id tipo Producto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "acti_usua",
                        "in": "path",
                        "description": "ingresar el usuario que va a registrar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdProyTarea": {
                                        "type": "string"
                                    },
                                    "intIdProy": {
                                        "type": "string"
                                    },
                                    "intIdTipoProducto": {
                                        "type": "string"
                                    },
                                    "varCodigoPaquete": {
                                        "type": "string"
                                    },
                                    "acti_usua": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProyTarea": "123",
                                    "intIdProy": "124",
                                    "intIdTipoProducto": "120",
                                    "varCodigoPaquete": "I002",
                                    "acti_usua": "usuarios"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exito."
                    },
                    "201": {
                        "description": "Si no existe el paquete entonces se agrega."
                    },
                    "408": {
                        "description": "Error diferentes paquetes."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
        "/GestionPartList/public/index.php/vali_proy_pque": {
            "post": {
                "tags": [
                    "Gestion Proyecto Paquete"
                ],
                "summary": "Validar El proyecto paquete",
                "operationId": "App\\Http\\Controllers\\ProyectopaqueteController::vali_proy_pque",
                "parameters": [
                    {
                        "name": "intIdProy",
                        "in": "path",
                        "description": "ingresar el id proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdTipoProducto",
                        "in": "path",
                        "description": "ingresar el id tipo producto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdProyTarea",
                        "in": "path",
                        "description": "ingresar el id proyecto tarea",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "varCodigoPaquete",
                        "in": "path",
                        "description": "ingresar codigo de paquete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdProyTarea": {
                                        "type": "string"
                                    },
                                    "intIdProy": {
                                        "type": "string"
                                    },
                                    "intIdTipoProducto": {
                                        "type": "string"
                                    },
                                    "varCodigoPaquete": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProy": "123",
                                    "intIdTipoProducto": "124",
                                    "intIdProyTarea": "120",
                                    "varCodigoPaquete": "I002",
                                    "acti_usua": "usuarios"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exito."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
        "/GestionPartList/public/index.php/obte_proy_paqu": {
            "post": {
                "tags": [
                    "Gestion Proyecto Paquete"
                ],
                "summary": "Obtener Proyecto Paquete",
                "operationId": "App\\Http\\Controllers\\ProyectopaqueteController::obte_proy_paqu",
                "parameters": [
                    {
                        "name": "varCodigoPaquete",
                        "in": "path",
                        "description": "ingresar el codigo elemento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdProy",
                        "in": "path",
                        "description": "ingresar el id proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdTipoProducto",
                        "in": "path",
                        "description": "ingresar el id tipo producto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdProy": {
                                        "type": "string"
                                    },
                                    "intIdTipoProducto": {
                                        "type": "string"
                                    },
                                    "intIdProyTarea": {
                                        "type": "string"
                                    },
                                    "varCodigoPaquete": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProy": "123",
                                    "intIdTipoProducto": "124",
                                    "intIdProyTarea": "120",
                                    "varCodigoPaquete": "I002",
                                    "acti_usua": "usuarios"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exito."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
    }
}